pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sparkrun"
version = "0.2.0"
description = "Launch and manage Docker-based inference workloads on NVIDIA DGX Spark systems"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "scitrera.ai", email = "open-source-team@scitrera.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Clustering",
]
dependencies = [
    "scitrera-app-framework==0.0.69",
    "vpd==0.9.13",
    "six==1.17.0", # transitive dependency of vpd but we pin the version for security reasons
    "click==8.3.1",
    "pyyaml==6.0.3",
    "huggingface-hub==1.8.0",
    "textual==8.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-cov==7.1.0",
    "pytest-asyncio==1.3.0",
]

[project.scripts]
sparkrun = "sparkrun.cli:main"

[project.entry-points."sparkrun.runtimes"]
vllm-distributed = "sparkrun.runtimes.vllm_distributed:VllmDistributedRuntime"
vllm-ray = "sparkrun.runtimes.vllm_ray:VllmRayRuntime"
sglang = "sparkrun.runtimes.sglang:SglangRuntime"
eugr-vllm = "sparkrun.runtimes.eugr_vllm_ray:EugrVllmRayRuntime"
llama-cpp = "sparkrun.runtimes.llama_cpp:LlamaCppRuntime"
trtllm = "sparkrun.runtimes.trtllm:TrtllmRuntime"

[project.entry-points."sparkrun.benchmarking"]
llama-benchy = "sparkrun.benchmarking.llama_benchy:LlamaBenchyFramework"

[project.entry-points."sparkrun.builders"]
docker-pull = "sparkrun.builders.docker_pull:DockerPullBuilder"
eugr = "sparkrun.builders.eugr:EugrBuilder"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"sparkrun.scripts" = ["*.sh"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 140
target-version = "py312"
